Manage npm dependencies (#2582)
* Remove npm package canvas

Canvas is used as an indirect dependency of jsdom and should not be
listed in the npm dependency section.

* Move npm dependencies into correct section

This commit moves "@storybook/client-api" and "postcss-import" to the
devDependencies section of the package file. These dependencies are not
referenced in runtime source code and only used in the test and build
process.