Increment required Elixir version
This patch increments the required Elixir compiler version to 1.15, to
be in sync with the `.tool-versions` file, used by `asdf`.

While this might not be strictly necessary, it might be worth to keep
the development version on the same version as the required version, so
confusion can be prevented.
The current version of SUSE SEL has Elixir 1.16 available to them.
https://packagehub.suse.com/packages/elixir/