Update installation instructions (#2594)
* Add version and additional requirements

This patch adds preferred versions to the list of required dependencies,
so the potential contributors will not need to reference the
`.tool-versions` file. This allows them to not be familiar or use `asdf`
and still contribute to the project.

* Document additional dependencies

During the setup on a new mac, it became apparent that some dependencies
are not necessarily pre-installed. This patch documents those
dependencies in a separate section and marks them as optional for most
users.

* Increment required Elixir version

This patch increments the required Elixir compiler version to 1.15, to
be in sync with the `.tool-versions` file, used by `asdf`.

While this might not be strictly necessary, it might be worth to keep
the development version on the same version as the required version, so
confusion can be prevented.
The current version of SUSE SEL has Elixir 1.16 available to them.
https://packagehub.suse.com/packages/elixir/

* Fix issues raised in the Review

This commit fixes issues that where raised by eugen.maksymenko@suse.com,
jamie.rodriguez@suse.com and xarbulu@suse.com. Thanks again!