Upgrade Node.js to version 20.14.0 LTS (#2660)
* Upgrade Node.js to version 20.14.0 LTS

This commit upgrade the Node.js version to 20.14.0.

I am running this for several weeks on my MacBook Air M2 and it works
without any problems. This will allow us to upgrade Sotorybook and
ESLint, once all our ESLint plugins are compatible with ESLint 9.

* Update Node.js in GitHub CI

This brings the CI up to the same version used in the development and
release environment.

* Update npm for the trento-web RPM

This commit replaces the npm version for the RPM, to use node v20's
version of npm.