Display affected systems to the user
This patch displays the systems affected by a advisory errata to the
user. It exposes the new backend API in the advisory errata page and
updates the relevant tests. This change required an update to the
factories.

Furthermore, I added a section for showcasing the "empty data" state in
the advisory details page.