Affected packages (#2763)
* Add factory for affected packages

* Enhance advisory errata endpoint with affected packages

This patch adds the affected packages to the advisory errata endpoint
and adds an openAPI spec for the changes.

* Update advisory errata backend tests for affected packages

This patch adds unit test cases and updates the existing cases to
include the new data available from the affected packages discovery.

* Update AdvisoryDetails component to display affected package data

This patch updates the AdvisoryDetails component to display the affected
packages.

* Update AdvisoryDetails frontend test and story to include affected packages

This patch updates and adds the frontend test and story to include the
affected packages data.

* Fix formatting

* Fix spelling

* Changes suggested by Alessio

* Expose affected systems in the errata details

This patch adds the affected systems from the discovery result to the
errata details response. Therefore, this patch contains an update to
the openAPI specification. Furthermore, it adds new and updates
existing tests to include the changes.

* Reword openAPI specification for affected packages

This patch rewords the openAPI specification for the affected packages.
I changed the words, to clarify the difference between a package
version and a release.

* Display affected systems to the user

This patch displays the systems affected by a advisory errata to the
user. It exposes the new backend API in the advisory errata page and
updates the relevant tests. This change required an update to the
factories.

Furthermore, I added a section for showcasing the "empty data" state in
the advisory details page.