Adust length of "Results per page"-field (#2968)
* Adust length of "Results per page"-field

This commit adjusts the length of the "Results per page"-field in the
views containing pagination.  The field now adjusts its size based on
its content.  For this, I use a margin on the chevron element, but
utilising CSS' after pseudo class would be better

* Show pagination statistics

This commit shows the current item range as well as the total amount of
items in the bottom of the table

* Fix Storybook

* Adjust spacing

* Fix Table empty state bug

The table should determine its empty state based on the filtered data,
because that can be less than the unfiltered data

* Add test and move component to a seperate file

* Move size limitation to pagination component

* Reset styles of Select

* pls :'''')