Introduces an additional test dep
ProcessTree helps look for keys in ancestors of a given process. This is useful for instance for scenarios where there is an async call/spawn to a child process from a parent process and one wishes to check the both the parent and child processes for a key in their process dictionaries. Relevant for tests involving SMLM settings.