Refactors dispatch functions and their usage (#3715)
* Refactor

...refactors and centralizes commanded dispatch related functions

* Fixes for credo

* Adds tests

* Refactors tests

* Address PR comment

Make commanded() an internal/private function and only expose a public dispatch function in its place.

* Addresses PR comment

Consolidates the public api of `CommandedUtils` by renaming `maybe_correlated_dispatch` function to `correlated_dispatch/2` and also makes the second parameter a required one (it was optional previously). This is because there are no scenarios when this function actually relies on the optional/implied-default parameter.